Changing functionality to accompany MSO24
We swapped oscilloscopes from the Tektronix TDS2022B to the Tektronix MSO24. Since the waveform output format is different, I have to adjust some things. Most of the 425 functionality is back, but I still have to reimplement the outputting of the final CSV once I decide on a better format to include the R_shunt and gain values.

I also have to implement the ability to read in the three channels. my plan is that the x-axis will be CH1, y-axis will be CH2, and z-axis will be CH3.